Rename utils/plurals.ts to utils/inflector.ts
