+++ packages/generator/src/index.ts
@@ -8,4 +8,4 @@ export * from "./generators/form-generator"
 export * from "./generator"
 export * from "./conflict-checker"
 export {getLatestVersion} from "./utils/get-latest-version"
-export {singleCamel, singlePascal, pluralCamel, pluralPascal} from "./utils/plurals"
+export {singleCamel, singlePascal, pluralCamel, pluralPascal} from "./utils/inflector"
