+++ packages/generator/src/utils/inflector.ts
